fixed download, changed Path to File

+++ src/main/java/torrent/client/protocol/Get.java
@@ -39,7 +39,7 @@ public final class Get {
 
         @Override
         public void write(DataOutputStream dataOutputStream) throws IOException {
-            dataOutputStream.writeInt(TAG);
+            dataOutputStream.writeByte(TAG);
             dataOutputStream.writeInt(id);
             dataOutputStream.writeInt(part);
         }
